Fixes #26 #27 #37 #40 #41 - Created models for gameflow.
Gameflow structures were created. Each user will have a GameFlowStatus associated to his/her GameFlow
GameFlow is assigned by the institutional program in which user is involved.
Different dependency attributed are associated to each game in the gameflow and to the gameflow as a whole.

Structures are created but no function is implemented. We will implement them as they are needed by views.

Gameflows are a set of game revisions. Different programs may have the same game with different versions running.
Then a game revision class was created based on OLPC impl.
Gameflow indicates which game and which game version are available
for playing. Then, a GameRevision model was creating for admin
versioning of games.