Fixes #55 - game_flow returns the list of games per skill level
Now the game_flow returns the list of games per skill level, with
a boolean attribute that shows if that game is enabled or not,
depending on game_flow rules.

Created two types of rules: finished game and partial completion
game dependencies.

This game list is returned from the gameflow view which needs
to be authenticated.
A login page, copied from admin site, was created for login a user.
Thus, we changed template directory configuration in local settings
file.

Fixes #56 - Test cases for enabled and disabled games were created.