Fixes #64 #70 #71 #73 - Created gameflow view
At home page, a "play" button was added to access the gameflow view.
If user is not logged in, login page is presented.
Changed login page to receive focus on the username textbox.
For this, a endscript block was added to the base template.

Once logged in, gameflow is presented. All games are presented separated
by their skills in a striped table. Also, a tooltip is presented
with the game explanations.

For each game, permission and requirements are checked and presented
accordingly, i.e. game, game locked or game not available for web.
For the web implementation, a web_gamenode class was created with
information related to game display in web platform, for instance tooltips.
Note that GameNode class was changed. Now each GameNode belongs only
to one GameFlow. GameNodes cannot be shared because of the rules dependencies.

Games are then accessed by the gameflow url indicating the gamenode to
play. If permissions are ok, then game is served. The game serving is still not
finished as templates for games must be designed.
In case, permissions are not enough, a HTML indicating the error is presented.

For each game, resources are accesed with the URL of the games:
games/game_name/game_version/ through the serve_ functions.
The allowed directories are: res, gamefiles and screenshots.
to identify uniquely a game by its name, we changed the field to slug format
with a unique contraint.