Fixes #65 #81 - Created page for listing and adding games.
Finished list of games, with modal dialogs for adding game,
game-revision or .zip packages for web implementation.

Added gamelist template which shows the current game of
Mate Marote. Add new game and add new revision buttons
are implemented with modal dialog.

Once game revision is created, upload button for web
implementation appears, also as a modal dialog.
If .zip file is uploaded, correct directories are
verified, and in case everything is OK, file is extracted.
For this purpose, a new class was created that encapsulates
functions for web implementation packades, e.g. directories.